+++ KikoleSite/Dtos/EntryDto.cs
@@ -2,8 +2,8 @@
 {
     public class EntryDto : EntryBaseDto
     {
-        public long Id { get; set; }
-        public long PlayerId { get; set; }
+        public uint Id { get; set; }
+        public uint PlayerId { get; set; }
         public bool IsSimulatedDate { get; set; }
     }
 }
