+++ KikoleSite/Models/RankingEntry.cs
@@ -11,18 +11,18 @@ namespace KikoleSite.Models
         private readonly Dictionary<Level, int> _levelPoints;
         private readonly Dictionary<Level, int> _levelUntiedRecordsCount;
         private readonly Dictionary<Level, int> _levelRecordsCount;
-        private readonly Dictionary<Level, long> _levelCumuledTime;
-        private readonly Dictionary<Stage, Dictionary<Level, (int, int, long?, System.DateTime?)>> _details;
+        private readonly Dictionary<Level, int> _levelCumuledTime;
+        private readonly Dictionary<Stage, Dictionary<Level, (int, int, int?, System.DateTime?)>> _details;
 
         public IReadOnlyDictionary<Level, int> LevelPoints => _levelPoints;
         public IReadOnlyDictionary<Level, int> LevelUntiedRecordsCount => _levelUntiedRecordsCount;
         public IReadOnlyDictionary<Level, int> LevelRecordsCount => _levelRecordsCount;
-        public IReadOnlyDictionary<Level, long> LevelCumuledTime => _levelCumuledTime;
-        public IReadOnlyDictionary<Stage, IReadOnlyDictionary<Level, (int, int, long?, System.DateTime?)>> Details => _details.ToDictionary(d => d.Key, d => (IReadOnlyDictionary<Level, (int, int, long?, System.DateTime?)>)d.Value);
+        public IReadOnlyDictionary<Level, int> LevelCumuledTime => _levelCumuledTime;
+        public IReadOnlyDictionary<Stage, IReadOnlyDictionary<Level, (int, int, int?, System.DateTime?)>> Details => _details.ToDictionary(d => d.Key, d => (IReadOnlyDictionary<Level, (int, int, int?, System.DateTime?)>)d.Value);
 
         internal RankingEntry(Game game, PlayerDto player) : base(game, player)
         {
-            _details = new Dictionary<Stage, Dictionary<Level, (int, int, long?, System.DateTime?)>>();
+            _details = new Dictionary<Stage, Dictionary<Level, (int, int, int?, System.DateTime?)>>();
 
             _levelPoints = ToLevelDictionary(0);
             _levelUntiedRecordsCount = ToLevelDictionary(0);
@@ -56,12 +56,12 @@ namespace KikoleSite.Models
             return points;
         }
 
-        private Dictionary<Level, (int, int, long?, System.DateTime?)> GetDetailsByLevel(Stage stage)
+        private Dictionary<Level, (int, int, int?, System.DateTime?)> GetDetailsByLevel(Stage stage)
         {
-            Dictionary<Level, (int, int, long?, System.DateTime?)> detailsByLevel;
+            Dictionary<Level, (int, int, int?, System.DateTime?)> detailsByLevel;
             if (!_details.ContainsKey(stage))
             {
-                detailsByLevel = new Dictionary<Level, (int, int, long?, System.DateTime?)>();
+                detailsByLevel = new Dictionary<Level, (int, int, int?, System.DateTime?)>();
                 _details.Add(stage, detailsByLevel);
             }
             else
