+++ KikoleSite/Providers/IIntegrationProvider.cs
@@ -9,7 +9,7 @@ namespace KikoleSite.Providers
     {
         Task<RefreshEntriesResult> RefreshAllEntriesAsync(Game game);
 
-        Task<RefreshEntriesResult> RefreshPlayerEntriesAsync(long playerId);
+        Task<RefreshEntriesResult> RefreshPlayerEntriesAsync(uint playerId);
 
         Task<RefreshEntriesResult> RefreshEntriesToDateAsync(DateTime stopAt);
